// Web/src/interfaces/Species.tsx
  export interface ResultsSpeciesEntity {
    name: string;
    classification: string;
    designation: string;
    average_height: string;
    skin_colors: string;
    hair_colors: string;
    eye_colors: string;
    average_lifespan: string;
    homeworld: string;
    language: string;
    people?: (string)[] | null;
    films?: (string)[] | null;
    created: string;
    edited: string;
    url: string;
  }